renamed _misc_functions to _functions.  Changed __init__ to import _functions.
