+++ __init__.py
@@ -28,7 +28,7 @@ import _plot             as plot        ; plot._settings = settings
 import _data             as data        ; data._settings = settings
 import _dialogs          as dialogs     ; dialogs._settings = settings
 import _pylab_tweaks     as tweaks      ; tweaks._settings = settings
-import _misc_functions   as fun         ; fun._settings = settings
+import _functions   as fun         ; fun._settings = settings
 
 plot.tweaks._pylab_colormap._settings = settings
